修正_total型別問題

+++ KafkaWebAPI/ViewModels/KafkaProducerDoctorViewModel.cs
@@ -32,7 +32,7 @@ namespace KafkaWebAPI.ViewModels
 
         public int Days { get; set; }
 
-        public int Total { get; set; }
+        public double Total { get; set; }
 
         public string? Remark { get; set; }
     }
